Clean up MarkDirty
